xlat: Use the TLBI of the correct EL for the context
Instead of executing the TLBI instruction that corresponds to the EL in
which the code is running, use the one that corresponds to the context.

This fixes the case in which a context of translation tables used in
EL1, but managed from EL3, would be incorrectly invalidated with
TLBI instructions for EL3 instead of EL1.

Change-Id: I7e1c4722dc4edace2109d3429c072c4c3375a9b6
Signed-off-by: Antonio Nino Diaz <antonio.ninodiaz@arm.com>