Add support for Secure Partition setup in BL31
This patch adds functions and data structures that will be used by the Secure
Partitions manager (in later patches) to initialise the architectural context
for execution of a Secure Partition in S-EL0.

A minimal S-EL1 exception vector table is implemented to relay SVC requests from
S-EL0 to the Secure Partition manager in EL3 as SMC requests.

Memory is allocated in the BL31 linker script for populating translation tables
for the S-EL1/EL0 regime from BL31.

A secure context in the context management library is initialised to enable
entry into the partition during cold boot.

Change-Id: I51ba74e5003c2ae19b5497ebb8fa2d4d09f04ad9
Signed-off-by: Achin Gupta <achin.gupta@arm.com>