SPM: Pass per-cpu shared buffer size to partition
This patch modifies the earlier approach of sending the total size of the shared
buffer between ARM TF and the Secure Partition as a part of the boot
information. The per-cpu size of the buffer is populated instead now. The
partition can use the cpu count to determine the total size of the shared
buffer.

Change-Id: Ifb5fdd86dd1c2b4697155ca42c2f9423bc1f16fa
Signed-off-by: Achin Gupta <achin.gupta@arm.com>